Add App and separate conteiner
